change route datall to panel-admin and other things accordingly
