+++ client/src/components/PanelAdmin/Register.tsx
@@ -5,36 +5,35 @@ import { boolean } from 'zod';
 
 import { useNavigate } from 'react-router-dom';
 
-const RegisterWithRole: React.FC = () => {
+const Register: React.FC = () => {
   const [fullName, setFullName] = useState<string>('');
   const [email, setEmail] = useState<string>('');
   const [password, setPassword] = useState<string>('');
   const [confirmPassword, setConfirmPassword] = useState<string>('');
-  const[role,setRole] =useState<string>('')
 
   const [fullNameError, setFullNameError] = useState<string>('');
   const [emailError, setEmailError] = useState<string>('');
   const [passwordError, setPasswordError] = useState<string>('');
   const [confirmPasswordError, setConfirmPasswordError] = useState<string>('');
-  const[roleError,setRoleError] =useState<string>('')
 
-  const [changeState, setChangeState] = useState<boolean>(false);
+  const [changeState ,setChangeState] = useState<boolean>(false)
 
-  // const registerPath =register()
-  const navigate = useNavigate();
+// const registerPath =register()
+const navigate= useNavigate()
 
-  useEffect(() => {
+
+useEffect(()=>{
     const fetchBalance = async () => {
-      try {
-        const response = await axios.get('http://localhost:3090/api/balance');
-        console.log('Balance:', response.data);
-      } catch (error) {
-        console.error('Error fetching balance:', error);
-      }
-    };
-
-    fetchBalance();
-  }, [changeState]);
+        try {
+          const response = await axios.get('http://localhost:3090/api/balance');
+          console.log('Balance:', response.data);
+        } catch (error) {
+          console.error('Error fetching balance:', error);
+        }
+      };
+  
+      fetchBalance(); 
+},[changeState])
 
   const fullNameHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
     setFullName(e.target.value);
@@ -72,15 +71,6 @@ const RegisterWithRole: React.FC = () => {
     }
   };
 
-  const roleHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setRole(e.target.value);
-    if (e.target.value !== 'ADMIN' && e.target.value !== 'USER') {
-      setRoleError('role is wrong.');
-    } else {
-      setRoleError('');
-    }
-  };
-
   const submitHandler = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     if (
@@ -91,8 +81,7 @@ const RegisterWithRole: React.FC = () => {
       fullNameError.length === 0 &&
       emailError.length === 0 &&
       passwordError.length === 0 &&
-      confirmPasswordError.length === 0 &&
-      roleError.length === 0
+      confirmPasswordError.length === 0
     ) {
       console.log('submit');
 
@@ -102,29 +91,31 @@ const RegisterWithRole: React.FC = () => {
         username: fullName,
         confirm_password: confirmPassword,
         password: password,
-        role: role
       };
 
-      const additionalData = { emailVerified: false };
+      const additionalData = {emailVerified :false}
 
       try {
-        const response = await axios.post('http://localhost:3090/api/auth/register', user);
+        const response = await axios.post('http://localhost:3090/api/auth/register', user );
         // const response = await axios.post(`http://localhost:3090${registerPath}`, user );
 
+        
         console.log('User Created:', response);
 
-        navigate('/datall/welcome', { state: user, replace: true });
+        navigate('/panel-admin/welcome', { state:user ,replace: true })
+
+        
       } catch (error) {
         console.error('Error creating user', error);
       }
-    } else {
-      setChangeState((prevState) => !prevState);
+    }else{
+        setChangeState(prevState=>!prevState)
+        
     }
   };
 
   return (
     <form className="flex flex-col gap-3" onSubmit={submitHandler}>
-        <h2>with roles:</h2>
       {/* fullName */}
       <div className="flex items-center justify-between">
         <label htmlFor="fullName">fullName: </label>
@@ -175,26 +166,12 @@ const RegisterWithRole: React.FC = () => {
       {confirmPasswordError.length > 0 && (
         <span className="text-red-500">{confirmPasswordError}</span>
       )}
-      {/* role */}
-      <div className="flex items-center justify-between">
-        <label htmlFor="role">role: </label>
-        <input
-          type="text"
-          placeholder="role"
-          id="role"
-          onChange={roleHandler}
-          className="rounded-2xl border px-3.5 py-2 focus:border-green-500 focus:outline-none"
-        />
-      </div>
-      {roleError.length > 0 && (
-        <span className="text-red-500">{roleError}</span>
-      )}
       <button type="submit" className="w-full rounded-2xl bg-green-500 py-2 text-white">
         Continue
       </button>
-      
     </form>
   );
-};
+}
 
-export default RegisterWithRole;
+
+export default Register
