+++ client/src/routes/Layouts/PanelAdmin.tsx
@@ -1,13 +1,13 @@
-import Register from '~/components/Datall/Register';
-import RegisterWithRole from '~/components/Datall/RegisterWithRole';
+import Register from '~/components/PanelAdmin/Register';
+import RegisterWithRole from '~/components/PanelAdmin/RegisterWithRole';
 import { Outlet, useLocation } from 'react-router-dom';
 import { Box, Container, Typography } from '@mui/material';
 
-const Datall: React.FC = () => {
+const PanelAdmin: React.FC = () => {
   const location = useLocation();
 
-  // return <>{location.pathname === '/datall/welcome' ? <Outlet/> : <Register /> }</>
-  // return <>{location.pathname === '/datall/welcome' ? <Outlet/> : <RegisterWithRole /> }</>
+  // return <>{location.pathname === '/panel-admin/welcome' ? <Outlet/> : <Register /> }</>
+  // return <>{location.pathname === '/panel-admin/welcome' ? <Outlet/> : <RegisterWithRole /> }</>
   // return <div className="flex justify-center"><Outlet/> <div className=" bg-red-400 "></div></div>
   return (
     <Container sx={{ bgcolor: 'white', height: '100vh' }}>
@@ -19,4 +19,4 @@ const Datall: React.FC = () => {
   );
 };
 
-export default Datall;
+export default PanelAdmin;
