+++ client/src/routes/index.tsx
@@ -16,10 +16,10 @@ import ShareRoute from './ShareRoute';
 import ChatRoute from './ChatRoute';
 import Search from './Search';
 import Root from './Root';
-import Datall from './Layouts/Datall';
-import WelcomeUser from '~/components/Datall/WelcomeUser';
-import AddBalance from '~/components/Datall/AddBalance';
-import RegisterWithRole from '~/components/Datall/RegisterWithRole';
+import PanelAdmin from './Layouts/PanelAdmin';
+import WelcomeUser from '~/components/PanelAdmin/WelcomeUser';
+import UserManagement from '~/components/PanelAdmin/UserManagement';
+import RegisterWithRole from '~/components/PanelAdmin/RegisterWithRole';
 
 const AuthLayout = () => (
   <AuthContextProvider>
@@ -54,16 +54,16 @@ export const router = createBrowserRouter([
     ],
   },
   {
-    path: 'datall',
-    element: <Datall />,
+    path: 'panel-admin',
+    element: <PanelAdmin /> ,
     children: [
       {
         path: 'welcome',
         element: <WelcomeUser />,
       },
       {
-        path: 'add-balance',
-        element: <AddBalance />,
+        path: 'user-management',
+        element: <UserManagement />,
       },
       {
         path: 'register-with-role',
